+++ analysis/import/SourceCodeParser/src/test/resources/max-nesting-level/lambda_function.java
@@ -1,0 +1,12 @@
+class LambdaFunction {
+
+    public void lamba() {
+        FuncInterface fobj = (int x) -> {
+            if(a<b) {
+                System.out.println(2 * x);
+            }
+        };
+
+        fobj.abstractFunction(5);
+    }
+}